adk_lab/requirements_cloudrun.txt
a2a-sdk
langchain
langgraph
langchain-core
uvicorn
google-cloud-logging